updated auth logic
