+++ app/page.tsx
@@ -1,6 +1,6 @@
 'use client'
 
-import { useState, useEffect } from 'react'
+import { useState } from 'react'
 import { useChat } from 'ai/react'
 
 import { Chat } from '@/components/chat'
@@ -10,14 +10,12 @@ import NavBar from '@/components/navbar'
 
 import { supabase } from './../lib/supabase'
 import { AuthDialog } from '@/components/AuthDialog'
-import { Session } from '@supabase/supabase-js'
-import { getUserAPIKeys } from '@/lib/apikeys'
+import { useAuth } from '@/lib/auth'
 
 export default function Home() {
   const [selectedTemplate, setSelectedTemplate] = useState(SandboxTemplate.CodeInterpreterMultilang)
-  const [session, setSession] = useState<Session | null>(null)
-  const [apiKey, setApiKey] = useState<string | undefined>(undefined)
   const [isAuthDialogOpen, setAuthDialog] = useState(false)
+  const { session, apiKey } = useAuth(setAuthDialog)
 
   const { messages, input, handleInputChange, handleSubmit, data } = useChat({
     api: '/api/chat',
@@ -33,30 +31,7 @@ export default function Home() {
   // Get the latest tool invocation
   const latestToolInvocation = latestMessageWithToolInvocation?.toolInvocations?.[0]
 
-  useEffect(() => {
-    supabase.auth.getSession().then(({ data: { session } }) => {
-      setSession(session)
-    })
-
-    const {
-      data: { subscription },
-    } = supabase.auth.onAuthStateChange((_event, session) => {
-      setSession(session)
-
-      if (_event === 'SIGNED_IN') {
-        setAuthDialog(false)
-        getUserAPIKeys(session as Session).then(setApiKey)
-      }
-
-      if (_event === 'SIGNED_OUT') {
-        setApiKey(undefined)
-      }
-    })
-
-    return () => subscription.unsubscribe()
-  }, [])
-
-  function handleSubmitAuth(e: React.FormEvent<HTMLFormElement>) {
+  function handleSubmitAuth (e: React.FormEvent<HTMLFormElement>) {
     if (!session) {
       e.preventDefault()
       return setAuthDialog(true)
@@ -65,10 +40,14 @@ export default function Home() {
     handleSubmit(e)
   }
 
+  function logout () {
+    supabase.auth.signOut()
+  }
+
   return (
     <main className="flex min-h-screen max-h-screen">
       <AuthDialog open={isAuthDialogOpen} setOpen={setAuthDialog} supabase={supabase} />
-      <NavBar session={session} showLogin={() => setAuthDialog(true)} />
+      <NavBar session={session} showLogin={() => setAuthDialog(true)} signOut={logout} />
         <div className="flex-1 flex space-x-8 w-full pt-16 pb-8 px-4">
           <Chat
             messages={messages}
