+++ lib/supabase.ts
@@ -1,2 +1,6 @@
 import { createClient } from '@supabase/supabase-js'
-export const supabase = createClient('https://ntjfcwpzsxugrykskdgi.supabase.co', 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJzdXBhYmFzZSIsInJlZiI6Im50amZjd3B6c3h1Z3J5a3NrZGdpIiwicm9sZSI6ImFub24iLCJpYXQiOjE2NzgxODIxNDEsImV4cCI6MTk5Mzc1ODE0MX0.3Xiy4rzoCfYxMFsyeoPgONGJZ2QNQtp-ZzyMU7K0kh0')
+
+export const supabase = createClient(
+  process.env.NEXT_PUBLIC_SUPABASE_URL!,
+  process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY!
+)
